app/requirements.txt
astroid==3.3.9
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dill==0.4.0
docker==7.1.0
Flask==3.1.0
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.7
pylint==3.3.7
requests==2.32.3
tomlkit==0.13.2
urllib3==2.4.0
Werkzeug==3.1.3
paho-mqtt==2.1.0
aiomqtt==2.4.0

